only if its a post

+++ packages/bsky/src/data-plane/server/indexing/plugins/post.ts
@@ -168,6 +168,7 @@ const insertFn = async (
       await db.insertInto('post_embed_external').values(externalEmbed).execute()
     } else if (isEmbedRecord(postEmbed)) {
       const { record } = postEmbed
+      const embedUri = new AtUri(record.uri)
       const recordEmbed = {
         postUri: uri.toString(),
         embedUri: record.uri,
@@ -176,36 +177,38 @@ const insertFn = async (
       embeds.push(recordEmbed)
       await db.insertInto('post_embed_record').values(recordEmbed).execute()
 
-      const quote = {
-        uri: uri.toString(),
-        cid: cid.toString(),
-        subject: record.uri,
-        subjectCid: record.cid,
-        createdAt: normalizeDatetimeAlways(obj.createdAt),
-        indexedAt: timestamp,
-      }
-      await db
-        .insertInto('quote')
-        .values(quote)
-        .onConflict((oc) => oc.doNothing())
-        .returningAll()
-        .executeTakeFirst()
+      if (embedUri.collection === lex.ids.AppBskyFeedPost) {
+        const quote = {
+          uri: uri.toString(),
+          cid: cid.toString(),
+          subject: record.uri,
+          subjectCid: record.cid,
+          createdAt: normalizeDatetimeAlways(obj.createdAt),
+          indexedAt: timestamp,
+        }
+        await db
+          .insertInto('quote')
+          .values(quote)
+          .onConflict((oc) => oc.doNothing())
+          .returningAll()
+          .executeTakeFirst()
 
-      const quoteCountQb = db
-        .insertInto('post_agg')
-        .values({
-          uri: record.uri.toString(),
-          quoteCount: db
-            .selectFrom('quote')
-            .where('quote.subjectCid', '=', record.cid.toString())
-            .select(countAll.as('count')),
-        })
-        .onConflict((oc) =>
-          oc
-            .column('uri')
-            .doUpdateSet({ quoteCount: excluded(db, 'quoteCount') }),
-        )
-      await quoteCountQb.execute()
+        const quoteCountQb = db
+          .insertInto('post_agg')
+          .values({
+            uri: record.uri.toString(),
+            quoteCount: db
+              .selectFrom('quote')
+              .where('quote.subjectCid', '=', record.cid.toString())
+              .select(countAll.as('count')),
+          })
+          .onConflict((oc) =>
+            oc
+              .column('uri')
+              .doUpdateSet({ quoteCount: excluded(db, 'quoteCount') }),
+          )
+        await quoteCountQb.execute()
+      }
     }
   }
 
@@ -365,24 +368,27 @@ const deleteFn = async (
     deletedEmbeds.push(deletedExternals)
   }
   if (deletedPosts) {
+    const embedUri = new AtUri(deletedPosts.embedUri)
     deletedEmbeds.push(deletedPosts)
 
-    await db.deleteFrom('quote').where('uri', '=', uriStr).execute()
-    await db
-      .insertInto('post_agg')
-      .values({
-        uri: deletedPosts.embedUri,
-        quoteCount: db
-          .selectFrom('quote')
-          .where('quote.subjectCid', '=', deletedPosts.embedCid)
-          .select(countAll.as('count')),
-      })
-      .onConflict((oc) =>
-        oc
-          .column('uri')
-          .doUpdateSet({ quoteCount: excluded(db, 'quoteCount') }),
-      )
-      .execute()
+    if (embedUri.collection === lex.ids.AppBskyFeedPost) {
+      await db.deleteFrom('quote').where('uri', '=', uriStr).execute()
+      await db
+        .insertInto('post_agg')
+        .values({
+          uri: deletedPosts.embedUri,
+          quoteCount: db
+            .selectFrom('quote')
+            .where('quote.subjectCid', '=', deletedPosts.embedCid)
+            .select(countAll.as('count')),
+        })
+        .onConflict((oc) =>
+          oc
+            .column('uri')
+            .doUpdateSet({ quoteCount: excluded(db, 'quoteCount') }),
+        )
+        .execute()
+    }
   }
   return deleted
     ? {
